remaining good changes from original coffeescript-generators branch
notable exceptions are that we don’t want to mess with coffeescript
and the base of the require js chain for reasons that have already
taken hours to debug and decide weren’t worth it

In other words, only pre-process with require js plugins after main.js
has been loaded. In the tests I’ve also decided that test/test-helper-*.js
should remain js for the time being as they may change in the future
and maintaining two versions at this point is superfluous.

Note that I still need to manually test using test/views/root.spec.coffee
as i don’t expect it will work at this point

Note that these are not cs tests for sub generators